updated price to be double

+++ app/src/main/java/com/example/user/shoppingcart/Shoppable.java
@@ -6,5 +6,5 @@ package com.example.user.shoppingcart;
 
 public interface Shoppable {
 
-    public int getPrice();
+    public double getPrice();
 }
